For app create with managed identity: refactor validator
