+++ src/spring-cloud/azext_spring_cloud/_params.py
@@ -24,9 +24,9 @@ from ._validators_enterprise import (only_support_enterprise, validate_builder_r
                                      validate_build_env, validate_target_module, validate_runtime_version)
 from ._app_validator import (fulfill_deployment_param, active_deployment_exist,
                              ensure_not_active_deployment, validate_deloy_path, validate_deloyment_create_path,
-                             validate_cpu, validate_memory, fulfill_deployment_param_or_warning,
-                             active_deployment_exist_or_warning, validate_create_app_with_user_identity_or_warning,
-                             validate_create_app_with_system_identity_or_warning)
+                             validate_cpu, validate_memory, fulfill_deployment_param_or_warning, active_deployment_exist_or_warning)
+from ._app_managed_identity_validator import (validate_create_app_with_user_identity_or_warning,
+                                              validate_create_app_with_system_identity_or_warning)
 from ._utils import ApiType
