+++ src/spring-cloud/azext_spring_cloud/commands.py
@@ -21,8 +21,8 @@ from ._transformers import (transform_spring_cloud_table_output,
                             transform_spring_cloud_gateway_output,
                             transform_api_portal_output)
 from ._validators_enterprise import (validate_gateway_update, validate_api_portal_update)
-from ._app_validator import (validate_app_identity_remove_or_warning,
-                             validate_app_identity_assign_or_warning)
+from ._app_managed_identity_validator import (validate_app_identity_remove_or_warning,
+                                              validate_app_identity_assign_or_warning)
 
 
 # pylint: disable=too-many-statements
